refactor : 데이터 패키지 생성 및 기존 클래스 패키지 이동
