+++ src/main/java/subway/domain/data/Station.java
@@ -1,9 +1,9 @@
-package subway.domain;
+package subway.domain.data;
 
-public class Line {
+public class Station {
     private String name;
 
-    public Line(String name) {
+    public Station(String name) {
         this.name = name;
     }
